# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bindizr HTTP API
  description: This is the API documentation for Bindizr
  contact:
    email: kevin136583@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.1.0-beta.7
paths:
...
  /external-dns/adjust:
    post:
      tags:
      - ExternalDNS
      summary: Canonicalize desired records without applying them
      description: 'Backs the adapter''s AdjustEndpoints step: returns each desired record in the canonical form applying it would store (uppercase type, sorted deduplicated presentation values), so external-dns compares desired state against the exact spelling GET /external-dns/records returns.'
      operationId: adjust_external_dns_records
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalDnsAdjustRequest'
        required: true
      responses:
        '200':
          description: Canonicalized records, in request order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalDnsAdjustResponse'
        '400':
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type, expected JSON request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /external-dns/changes:
    post:
      tags:
      - ExternalDNS
      summary: Apply an ExternalDNS change set atomically
      description: Applies creates, updates, and deletes in one transaction. Idempotent operations resolve to no change; only zones with a remaining delta advance their serial (once per request) and record IXFR history.
      operationId: apply_external_dns_changes
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalDnsChangesRequest'
        required: true
      responses:
        '200':
          description: Change set applied (or resolved to a no-op)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalDnsChangesResponse'
        '400':
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: The token is not allowed to manage a target record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No authoritative zone for a record name, or the token cannot see it
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Record conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '415':
          description: Unsupported media type, expected JSON request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /external-dns/domains:
    get:
      tags:
      - ExternalDNS
      summary: List the names ExternalDNS may manage
      description: 'The ExternalDNS domain filter the calling token''s grants come to: every zone name for a global token, otherwise one name per writable grant — the zone where the grant covers every name, the granted subtree where it does not. Each entry covers itself and everything under it, so a grant narrowed by record type or to one exact name reads wider here than it is.'
      operationId: list_external_dns_domains
      responses:
        '200':
          description: Manageable names
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalDnsDomainsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /external-dns/records:
    get:
      tags:
      - ExternalDNS
      summary: List the records of every ExternalDNS-managed zone
      description: 'Records of every zone the calling token may manage, restricted to the supported record types (A, AAAA, CNAME, TXT): one record per name and type, with absolute owner names and sorted presentation-form values.'
      operationId: list_external_dns_records
      responses:
        '200':
          description: Records of the allowed zones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalDnsRecordsResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: 'Generic error response: a plain description plus a machine-readable code.'
      required:
      - error
      - code
      properties:
        code:
          type: string
          example: ZONE_NOT_FOUND
        error:
          type: string
          example: Zone with name 'example.com' not found
    ExternalDnsAdjustRequest:
      type: object
      description: Request body for canonicalizing desired records without applying them.
      required:
      - records
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/ExternalDnsRecord'
    ExternalDnsAdjustResponse:
      type: object
      description: The request's records in the canonical form applying them would store.
      required:
      - records
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/ExternalDnsRecord'
    ExternalDnsChangesRequest:
      type: object
      description: Request body for applying an ExternalDNS change set atomically.
      properties:
        creates:
          type: array
          items:
            $ref: '#/components/schemas/ExternalDnsRecord'
        deletes:
          type: array
          items:
            $ref: '#/components/schemas/ExternalDnsRecord'
        updates:
          type: array
          items:
            $ref: '#/components/schemas/ExternalDnsRecordUpdate'
    ExternalDnsChangesResponse:
      type: object
      description: Summary of an applied ExternalDNS change set.
      required:
      - changed_zones
      - records_added
      - records_deleted
      properties:
        changed_zones:
          type: array
          items:
            type: string
          description: Zones whose serial advanced; empty when the request was a no-op.
          example:
          - example.com
        records_added:
          type: integer
          format: int32
          description: Counted one per value written, not per name and type.
          example: 2
          minimum: 0
        records_deleted:
          type: integer
          format: int32
          description: Counted one per value removed, not per name and type.
          example: 1
          minimum: 0
    ExternalDnsDomainsResponse:
      type: object
      description: |-
        The names the ExternalDNS caller may manage under its grants, each
        covering itself and everything under it.
      required:
      - domains
      properties:
        domains:
          type: array
          items:
            type: string
          example:
          - example.com
    ExternalDnsRecord:
      type: object
      description: |-
        One record of the ExternalDNS API: every value of one name and type. Names
        are absolute; TXT values are quoted presentation strings.
      required:
      - name
      - record_type
      - values
      properties:
        name:
          type: string
          example: app.example.com
        record_type:
          type: string
          example: A
        ttl:
          type:
          - integer
          - 'null'
          format: int32
          description: Optional on writes; an omitted or zero TTL resolves to the zone TTL.
          example: 300
        values:
          type: array
          items:
            type: string
          example:
          - 192.0.2.10
    ExternalDnsRecordUpdate:
      type: object
      description: 'A record replacement: `old` values are removed and `new` written in place.'
      required:
      - old
      - new
      properties:
        new:
          $ref: '#/components/schemas/ExternalDnsRecord'
        old:
          $ref: '#/components/schemas/ExternalDnsRecord'
    ExternalDnsRecordsResponse:
      type: object
      description: |-
        Records of every ExternalDNS-managed zone, one per name and type, in a
        deterministic order.
      required:
      - records
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/ExternalDnsRecord'
...
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
security:
- bearer_auth: []
...